+++ react-onsenui/src/components/Tab.jsx
@@ -1,5 +1,7 @@
-import SimpleWrapper from './SimpleWrapper.jsx';
+import PropTypes from 'prop-types';
 import 'onsenui/esm/elements/ons-tab';
+
+import onsCustomElement from './onsCustomElement';
 
 /**
  * @original ons-tab
@@ -14,10 +16,44 @@ import 'onsenui/esm/elements/ons-tab';
  *   Home
  * </Tap>
  */
-class Tab extends SimpleWrapper {
-  _getDomNodeName() {
-    return 'ons-tab';
-  }
-}
+const Tab = onsCustomElement('ons-tab');
+
+Tab.propTypes = {
+  /**
+   * @name icon
+   * @type string
+   * @description
+   *  [en]The icon name for the tab. Can specify the same icon name as <ons-icon>.[/en]
+   *  [ja][/ja]
+   */
+  icon: PropTypes.string,
+
+  /**
+   * @name activeIcon
+   * @type string
+   * @description
+   *  [en]The name of the icon when the tab is active.[/en]
+   *  [ja][/ja]
+   */
+  activeIcon: PropTypes.string,
+
+  /**
+   * @name label
+   * @type string
+   * @description
+   *  [en]The label of the tab item.[/en]
+   *  [ja][/ja]
+   */
+  label: PropTypes.string,
+
+  /**
+   * @name badge
+   * @type string
+   * @description
+   *  [en]Display a notification badge on top of the tab.[/en]
+   *  [ja][/ja]
+   */
+  badge: PropTypes.string
+};
 
 export default Tab;